requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pyTelegramBotAPI==3.6.6
requests==2.21.0
six==1.12.0
urllib3==1.24.2
python-telegram-bot==12.8
investpy==1.0
pandas==1.1.3
scipy==1.5.3
numpy==1.19.0
pyportfolioopt==1.2.6
matplotlib==3.3.2